fix(batch): validate atomic update operations in list order (#458)
Review on #560 found that the duplicate pre-pass could outrank the
per-operation required-field errors: a duplicate with nil Updates was
reported as a duplicate, and an earlier operation with an empty schema
was skipped so a later duplicate was reported first. Fold the required
field checks and duplicate detection into one list-order pre-pass
(validateAtomicUpdateOperations) so the lowest-index error always wins
and every check still runs before registry lookup and repository I/O.
The per-op checks leave the main loop since the pre-pass now owns them.

Move the batch write types out of types.go into types_batch.go: the file
was already over the 500-line limit on main and the new contract comment
pushed it further. The #458 tests move to their own file for the same
reason, and the mock repository counts reads and atomic writes so the
"before any read or write" guarantee is asserted directly.
